Bunch of exception handling for empty inputs and CRUD protection for against non admins

+++ 5Semester/5Semester/Controllers/UsersController.cs
@@ -11,12 +11,17 @@ using Microsoft.AspNetCore.Authentication;
 using System.Security.Cryptography;
 using System.Text;
 using Microsoft.AspNetCore.Cryptography.KeyDerivation;
+using MailKit.Net.Smtp;
+using MimeKit;
+using System.Net.Mail;
+using MimeKit.Text;
 
 namespace _5Semester.Controllers
 {
     public class UsersController : Controller
     {    
         private readonly _5SemesterContext _context;
+
         //const string SessionName = "_Name";
         //const string SessionAge = "_Age";
 
@@ -29,6 +34,65 @@ namespace _5Semester.Controllers
         {
             return View();
         }
+        public IActionResult PasswordRecoveryPage()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        public IActionResult PasswordRecoveryBtn(string email)
+        {
+            //if (ModelState.IsValid)
+            //{
+            //    try
+            //    {
+            //        //instantiate a new MimeMessage
+            //        var message = new MimeMessage();
+            //        //Setting the To e-mail address
+            //        message.To.Add(new MailboxAddress("E-mail Recipient Name", "recipient@domail.com"));
+            //        //Setting the From e-mail address
+            //        message.From.Add(new MailboxAddress("E-mail From Name", "from@domain.com"));
+            //        //E-mail subject 
+            //        message.Subject = contactViewModel.Subject;
+            //        //E-mail message body
+            //        message.Body = new TextPart(TextFormat.Html)
+            //        {
+            //            Text = contactViewModel.Message + " Message was sent by: " + contactViewModel.Name + " E-mail: " + contactViewModel.Email
+            //        };
+
+            //        //Configure the e-mail
+            //        using (var emailClient = new MailKit.Net.Smtp.SmtpClient())
+            //        {
+            //            emailClient.Connect("smtp.sparkpostmail.com", 587, false);
+            //            emailClient.Authenticate("SMTP_Injection", "76289861fbff7abd93a583e3aeeb5b5bb02e5ce8");
+            //            emailClient.Send(message);
+            //            emailClient.Disconnect(true);
+            //        }
+            //    }
+            //    catch (Exception ex)
+            //    {
+            //        ModelState.Clear();
+            //        ViewBag.Message = $" Oops! We have a problem here {ex.Message}";
+            //    }
+            //}
+
+            User userinfo = _context.User
+                         .Where(b => b.Email == email)
+                    .FirstOrDefault();
+            try
+            {
+                    if (email == userinfo.Email)
+                    {
+                    
+                    }
+            }
+            catch(NullReferenceException e)
+            {
+                
+            }
+                return View();
+        }
+
         public async Task<IActionResult> Logout()
         {
             HttpContext.Session.Clear();
@@ -90,19 +154,24 @@ namespace _5Semester.Controllers
         // GET: Users/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            string status = HttpContext.Session.GetString("sessionStatus");
+            if (status == "Admin")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.UserId == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
+                var user = await _context.User
+                    .FirstOrDefaultAsync(m => m.UserId == id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
 
-            return View(user);
+                return View(user);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Users/Create
@@ -116,7 +185,7 @@ namespace _5Semester.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("UserId,Username,Status,Password,DisplayName,Salt")] User user)
+        public async Task<IActionResult> Create([Bind("UserId,Username,Status,Password,DisplayName,Salt,Email")] User user)
         {
             if (ModelState.IsValid)
             {
@@ -172,7 +241,10 @@ namespace _5Semester.Controllers
         
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            string status = HttpContext.Session.GetString("sessionStatus");
+            if (status == "Admin")
+            {
+                if (id == null)
             {
                 return NotFound();
             }
@@ -183,6 +255,8 @@ namespace _5Semester.Controllers
                 return NotFound();
             }
             return View(user);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Users/Edit/5
@@ -192,7 +266,10 @@ namespace _5Semester.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("UserId,Username,Password,DisplayName")] User user)
         {
-            if (id != user.UserId)
+            string status = HttpContext.Session.GetString("sessionStatus");
+            if (status == "Admin")
+            {
+                if (id != user.UserId)
             {
                 return NotFound();
             }
@@ -219,24 +296,31 @@ namespace _5Semester.Controllers
                 return RedirectToAction(nameof(Index));
             }
             return View(user);
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         // GET: Users/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            string status = HttpContext.Session.GetString("sessionStatus");
+            if (status == "Admin")
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var user = await _context.User
-                .FirstOrDefaultAsync(m => m.UserId == id);
-            if (user == null)
-            {
-                return NotFound();
-            }
+                var user = await _context.User
+                    .FirstOrDefaultAsync(m => m.UserId == id);
+                if (user == null)
+                {
+                    return NotFound();
+                }
 
-            return View(user);
+                return View(user);
+                }
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Users/Delete/5
@@ -244,10 +328,14 @@ namespace _5Semester.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var user = await _context.User.FindAsync(id);
-            _context.User.Remove(user);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            string status = HttpContext.Session.GetString("sessionStatus");
+            if (status == "Admin") {
+                    var user = await _context.User.FindAsync(id);
+                _context.User.Remove(user);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction("Index", "Home");
         }
 
         private bool UserExists(int id)
